stack broken but input correct
stack module was changed as i removed the clock from it as it didn't seem necessary. Need to perform tests to see if it was vital to how it works. I am just worried about logic gate delays messing up the clock, so i am trying to remove the clock from as many regions as possible. Currently, inputs to stack are all correct. However,